Move national applicability to module
Allow the feature to be shared between formats
